create subscriptions table
