+++ src/main/resources/liquibase/2.8.x.x/base.xml
@@ -20,5 +20,6 @@
   <include relativeToChangelogFile="true" file="loinc.xml" />
   <include relativeToChangelogFile="true" file="notifications.xml" />
   <include relativeToChangelogFile="true" file="sample_program_audit_trail.xml" />
+  <include relativeToChangelogFile="true" file="notification_subscriptions.xml" />
 
 </databaseChangeLog>
