+++ src/main/resources/liquibase/2.8.x.x/notification_subscriptions.xml
